testing out different layers and testing suspicions
looks like resnet34 are more washed out, and I suspect that it's because I am not normalizing the input image. however, I'm having a problem normalizing the input - not sure why, but I cannot just put the normalization class as the first item in resnet34 model. I also tried normalizing the input image outside the model, but that did not work.